Clean up canonical tests for new scheme
